refactor: add constraint for workspace id field
