+++ src/core/features/attachments/AttachmentsController.test.ts
@@ -1,20 +1,19 @@
-import { makeAutoClosedDB } from 'src/__tests__/utils/makeAutoClosedDB';
-import { getUUID } from 'src/__tests__/utils/uuid';
+import { makeAppContext } from 'src/__tests__/utils/makeAppContext';
+import { makeAutoClosedSQLiteDB } from 'src/__tests__/utils/makeAutoClosedSQLiteDB';
 
 import { FilesController } from '../files/FilesController';
 import { InMemoryFS } from '../files/InMemoryFS';
 import { NotesController } from '../notes/controller/NotesController';
 import { AttachmentsController } from './AttachmentsController';
 
-const { getDB } = makeAutoClosedDB();
+const { getDB } = makeAutoClosedSQLiteDB();
+const getAppContext = makeAppContext(getDB);
 
 test('basic usage', async () => {
-	const db = await getDB();
+	const { db, workspaceId } = getAppContext();
 
-	const FAKE_WORKSPACE_ID = getUUID();
-
-	const notes = new NotesController(db, FAKE_WORKSPACE_ID);
-	const files = new FilesController(db, new InMemoryFS(), FAKE_WORKSPACE_ID);
+	const notes = new NotesController(db, workspaceId);
+	const files = new FilesController(db, new InMemoryFS(), workspaceId);
 
 	const [NOTE1, NOTE2] = await Promise.all([
 		notes.add({ title: '', text: '' }),
@@ -26,7 +25,7 @@ test('basic usage', async () => {
 		files.add(new File([], 'filename')),
 	]);
 
-	const attachments = new AttachmentsController(db, FAKE_WORKSPACE_ID);
+	const attachments = new AttachmentsController(db, workspaceId);
 	await attachments.set(NOTE1, [FILE_1, FILE_2]);
 	await attachments.set(NOTE2, [FILE_1, FILE_2, FILE_3]);
