+++ src/core/features/files/FilesController.test.ts
@@ -1,19 +1,18 @@
 // @vitest-environment jsdom
 
+import { makeAppContext } from 'src/__tests__/utils/makeAppContext';
 import { makeAutoClosedDB } from 'src/__tests__/utils/makeAutoClosedDB';
-import { getUUID } from 'src/__tests__/utils/uuid';
 
 import { createFileManagerMock } from './__tests__/mocks/createFileManagerMock';
 import { FilesController } from './FilesController';
 
-const FAKE_WORKSPACE_ID = getUUID();
-
 const { getDB } = makeAutoClosedDB();
+const getAppContext = makeAppContext(getDB);
 const fileManager = createFileManagerMock();
 
 test('Upload files', async () => {
-	const db = await getDB();
-	const files = new FilesController(db, fileManager, FAKE_WORKSPACE_ID);
+	const { db, workspaceId } = getAppContext();
+	const files = new FilesController(db, fileManager, workspaceId);
 
 	await files.add(new File(['Hello world'], 'hello.md', { type: 'text/markdown' }));
 	await files.add(new File(['foo'], 'foo.txt', { type: 'text/plain' }));
@@ -24,8 +23,8 @@ test('Upload files', async () => {
 });
 
 test('Fetch files', async () => {
-	const db = await getDB();
-	const files = new FilesController(db, fileManager, FAKE_WORKSPACE_ID);
+	const { db, workspaceId } = getAppContext();
+	const files = new FilesController(db, fileManager, workspaceId);
 
 	const filesList = await files.query();
 	expect(filesList).toEqual([
@@ -59,8 +58,8 @@ test('Fetch files', async () => {
 });
 
 test('Delete files', async () => {
-	const db = await getDB();
-	const files = new FilesController(db, fileManager, FAKE_WORKSPACE_ID);
+	const { db, workspaceId } = getAppContext();
+	const files = new FilesController(db, fileManager, workspaceId);
 
 	const filesList = await files.query();
 	expect(filesList).toHaveLength(4);
@@ -72,8 +71,8 @@ test('Delete files', async () => {
 });
 
 test('Fetch for non exists file in FS returns null', async () => {
-	const db = await getDB();
-	const files = new FilesController(db, fileManager, FAKE_WORKSPACE_ID);
+	const { db, workspaceId } = getAppContext();
+	const files = new FilesController(db, fileManager, workspaceId);
 
 	const filesList = await files.query();
