+++ src/core/features/notes/bin/DeletedNotesController.test.ts
@@ -1,5 +1,6 @@
 import ms from 'ms';
-import { getUUID } from 'src/__tests__/utils/uuid';
+import { createWorkspaceId, makeAppContext } from 'src/__tests__/utils/makeAppContext';
+import { makeAutoClosedSQLiteDB } from 'src/__tests__/utils/makeAutoClosedSQLiteDB';
 import { Mock } from 'vitest';
 import { openSQLite } from '@core/database/sqlite/openSQLite';
 import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
@@ -7,9 +8,9 @@ import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
 import { NotesController } from '../controller/NotesController';
 import { DeletedNotesController } from './DeletedNotesController';
 
-const FAKE_WORKSPACE_ID = getUUID();
-
 const dbFile = createFileControllerMock();
+const { getDB } = makeAutoClosedSQLiteDB({ file: dbFile });
+const getAppContext = makeAppContext(getDB);
 
 vi.useFakeTimers();
 
@@ -25,8 +26,8 @@ const getNoteIds = async (notes: NotesController) =>
 
 describe('Purge expired notes', () => {
 	test('Purge notes in bin that is deleted longer than retention time', async () => {
-		const db = await openSQLite(dbFile);
-		const notesController = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const notesController = new NotesController(db, workspaceId);
 		const bin = new DeletedNotesController(
 			{ notes: notesController },
 			{ retentionTime: ms('10d') },
@@ -78,8 +79,8 @@ describe('Purge expired notes', () => {
 	});
 
 	test('Modification time is considered when required', async () => {
-		const db = await openSQLite(dbFile);
-		const notesController = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const notesController = new NotesController(db, workspaceId);
 		const bin = new DeletedNotesController(
 			{ notes: notesController },
 			{ retentionTime: ms('10d'), considerModificationTime: true },
@@ -126,8 +127,8 @@ describe('Purge expired notes', () => {
 });
 
 test('Empty bin', async () => {
-	const db = await openSQLite(dbFile);
-	const notesController = new NotesController(db, FAKE_WORKSPACE_ID);
+	const { db, workspaceId } = getAppContext();
+	const notesController = new NotesController(db, workspaceId);
 	const bin = new DeletedNotesController(
 		{ notes: notesController },
 		{ retentionTime: ms('10d'), considerModificationTime: true },
@@ -184,7 +185,8 @@ describe('Service must delete expired notes', () => {
 
 		const dbFile = createFileControllerMock();
 		const db = await openSQLite(dbFile);
-		const notesController = new NotesController(db, FAKE_WORKSPACE_ID);
+		const workspaceId = await createWorkspaceId(db);
+		const notesController = new NotesController(db, workspaceId);
 		const bin = new DeletedNotesController(
 			{ notes: notesController },
 			{ retentionTime: ms('1d'), considerModificationTime: true },
@@ -268,7 +270,8 @@ describe('Service must delete expired notes', () => {
 
 		const dbFile = createFileControllerMock();
 		const db = await openSQLite(dbFile);
-		const notesController = new NotesController(db, FAKE_WORKSPACE_ID);
+		const workspaceId = await createWorkspaceId(db);
+		const notesController = new NotesController(db, workspaceId);
 		const bin = new DeletedNotesController(
 			{ notes: notesController },
 			{ retentionTime: ms('5y'), considerModificationTime: true },
@@ -315,7 +318,8 @@ describe('Service must delete expired notes', () => {
 
 		const dbFile = createFileControllerMock();
 		const db = await openSQLite(dbFile);
-		const notesController = new NotesController(db, FAKE_WORKSPACE_ID);
+		const workspaceId = await createWorkspaceId(db);
+		const notesController = new NotesController(db, workspaceId);
 		const bin = new DeletedNotesController(
 			{ notes: notesController },
 			{ retentionTime: ms('1d'), considerModificationTime: true },
