+++ src/core/features/notes/controller/NotesController.filters.test.ts
@@ -1,10 +1,10 @@
-import { getUUID } from 'src/__tests__/utils/uuid';
+import { createWorkspaceId, makeAppContext } from 'src/__tests__/utils/makeAppContext';
+import { makeAutoClosedSQLiteDB } from 'src/__tests__/utils/makeAutoClosedSQLiteDB';
 import { openSQLite } from '@core/database/sqlite/openSQLite';
+import { TagsController } from '@core/features/tags/controller/TagsController';
 import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
 
 import { NotesController } from './NotesController';
-
-const FAKE_WORKSPACE_ID = getUUID();
 
 vi.useFakeTimers();
 
@@ -12,7 +12,8 @@ test('filter by update time', async () => {
 	const db = await openSQLite(createFileControllerMock());
 	onTestFinished(db.close);
 
-	const notes = new NotesController(db, FAKE_WORKSPACE_ID);
+	const workspaceId = await createWorkspaceId(db);
+	const notes = new NotesController(db, workspaceId);
 
 	vi.setSystemTime('01/01/2001 12:00');
 	const note1 = await notes.add({ title: '2001', text: 'Dummy text' });
@@ -58,4 +59,255 @@ test('filter by update time', async () => {
 	).resolves.toEqual([
 		expect.objectContaining({ content: expect.objectContaining({ title: '2005' }) }),
 	]);
+});
+
+describe('data fetching', () => {
+	const { getDB } = makeAutoClosedSQLiteDB();
+	const getAppContext = makeAppContext(getDB);
+
+	const notesSample = Array(300)
+		.fill(null)
+		.map((_, idx) => {
+			return {
+				title: 'Note title #' + idx,
+				text: 'Note text #' + idx,
+			};
+		});
+
+	test('insert sample entries', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+
+		const ids: string[] = [];
+		for (const note of notesSample) {
+			ids.push(await registry.add(note));
+		}
+
+		const tags = new TagsController(db, workspaceId);
+		await tags.setAttachedTags(ids[0], [await tags.add('foo', null)]);
+		await tags.setAttachedTags(ids[1], [await tags.add('bar', null)]);
+	});
+
+	test('filter by tags', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const tags = new TagsController(db, workspaceId);
+
+		const tagsList = await tags.getTags();
+
+		await expect(
+			registry.get({
+				tags: [tagsList.find((tag) => tag.resolvedName === 'foo')!.id],
+			}),
+		).resolves.toHaveLength(1);
+
+		await expect(
+			registry.get({
+				tags: [tagsList.find((tag) => tag.resolvedName === 'bar')!.id],
+			}),
+		).resolves.toHaveLength(1);
+	});
+
+	test('method query consider filters', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const tags = new TagsController(db, workspaceId);
+
+		await expect(registry.query()).resolves.toHaveLength(300);
+		await expect(registry.query({ limit: 100 })).resolves.toHaveLength(100);
+		await expect(
+			registry.query({ limit: 100, meta: { isDeleted: true } }),
+		).resolves.toHaveLength(0);
+
+		const tagsList = await tags.getTags();
+		const fooTag = tagsList.find((tag) => tag.resolvedName === 'foo');
+		expect(fooTag).toBeDefined();
+
+		await expect(
+			registry.query({
+				tags: [fooTag!.id],
+			}),
+		).resolves.toHaveLength(1);
+	});
+
+	test('filter by tags and the deleted status', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const tags = new TagsController(db, workspaceId);
+
+		const tagsList = await tags.getTags();
+
+		const barTag = tagsList.find((tag) => tag.resolvedName === 'bar')!.id;
+		const [{ id: noteId }] = await registry.get({
+			limit: 1,
+			tags: [barTag],
+		});
+
+		// set deleted status
+		await registry.updateMeta([noteId], { isDeleted: true });
+		await expect(
+			registry.get({
+				tags: [barTag],
+				meta: { isDeleted: true },
+			}),
+		).resolves.toHaveLength(1);
+
+		// reset deleted status
+		await registry.updateMeta([noteId], { isDeleted: false });
+		await expect(
+			registry.get({
+				tags: [barTag],
+				meta: { isDeleted: true },
+			}),
+		).resolves.toHaveLength(0);
+	});
+
+	test('get entries by deletion status', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+
+		const notesId = await registry
+			.get({ limit: 10 })
+			.then((notes) => notes.map((note) => note.id));
+
+		// update status for 10 notes
+		vi.setSystemTime('2020-01-01T15:00:00.000Z');
+		await registry.updateMeta(notesId, { isDeleted: true });
+		await expect(registry.get({ meta: { isDeleted: true } })).resolves.toHaveLength(
+			10,
+		);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					from: new Date('2020-01-01T15:00:00.000Z'),
+					to: new Date('2020-01-01T15:00:00.000Z'),
+				},
+			}),
+			'Must find by exact time',
+		).resolves.toHaveLength(10);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					from: new Date('2020-01-01T15:00:00.000Z'),
+				},
+			}),
+			'Must find by exact "from" time',
+		).resolves.toHaveLength(10);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					from: new Date('2020-01-01T14:00:00.000Z'),
+				},
+			}),
+			'Must find by "from" time before deletion',
+		).resolves.toHaveLength(10);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					from: new Date('2020-01-01T15:00:00.001Z'),
+				},
+			}),
+			'Must not find by "from" time after deletion',
+		).resolves.toHaveLength(0);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					to: new Date('2020-01-01T15:00:00.000Z'),
+				},
+			}),
+			'Must find by exact "to" time',
+		).resolves.toHaveLength(10);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					to: new Date('2020-01-01T15:00:00.001Z'),
+				},
+			}),
+			'Must find by "to" time after deletion',
+		).resolves.toHaveLength(10);
+
+		await expect(
+			registry.get({
+				deletedAt: {
+					to: new Date('2020-01-01T14:00:00.000Z'),
+				},
+			}),
+			'Must not find by "to" time before deletion',
+		).resolves.toHaveLength(0);
+
+		// check only not deleted notes
+		await expect(registry.get({ meta: { isDeleted: false } })).resolves.toHaveLength(
+			notesSample.length - 10,
+		);
+	});
+
+	test('filter notes by archived status', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+
+		const notesId = await registry
+			.get({ limit: 30 })
+			.then((notes) => notes.map((note) => note.id));
+
+		await registry.updateMeta(notesId, { isArchived: true });
+		await expect(registry.get({ meta: { isArchived: true } })).resolves.toHaveLength(
+			30,
+		);
+		await expect(registry.get({ meta: { isArchived: false } })).resolves.toHaveLength(
+			notesSample.length - 30,
+		);
+	});
+
+	test('filters notes by bookmarks', async () => {
+		const { db, workspaceId } = getAppContext();
+
+		const registry = new NotesController(db, workspaceId);
+
+		const notesId = await registry
+			.get({ limit: 40 })
+			.then((notes) => notes.map((note) => note.id));
+
+		await registry.updateMeta(notesId, { isBookmarked: true });
+		await expect(
+			registry.get({ meta: { isBookmarked: true } }),
+		).resolves.toHaveLength(40);
+		await expect(
+			registry.get({ meta: { isBookmarked: false } }),
+		).resolves.toHaveLength(notesSample.length - 40);
+	});
+
+	test('method getLength consider filters', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+
+		const notesId = await registry
+			.get({ limit: 10 })
+			.then((notes) => notes.map((note) => note.id));
+		await registry.updateMeta(notesId, { isVisible: false });
+
+		await expect(registry.getLength({ meta: { isVisible: false } })).resolves.toBe(
+			10,
+		);
+		await expect(registry.getLength({ meta: { isVisible: true } })).resolves.toBe(
+			notesSample.length - 10,
+		);
+	});
+
+	test('method getById returns notes in the requested order', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+
+		const notes = await registry.get({ limit: 10 });
+
+		const shuffledNotes = notes.sort(() => Math.random() - 0.5);
+		const shuffledIds = shuffledNotes.map((n) => n.id);
+
+		await expect(registry.getById(shuffledIds)).resolves.toEqual(shuffledNotes);
+	});
 });
