+++ src/core/features/notes/controller/NotesController.test.ts
@@ -1,26 +1,18 @@
-/* eslint-disable @cspell/spellchecker */
+import { createWorkspaceId, makeAppContext } from 'src/__tests__/utils/makeAppContext';
 import { makeAutoClosedSQLiteDB } from 'src/__tests__/utils/makeAutoClosedSQLiteDB';
-import { getUUID } from 'src/__tests__/utils/uuid';
-import z from 'zod';
 import { openSQLite } from '@core/database/sqlite/openSQLite';
-import { InMemoryFS } from '@core/features/files/InMemoryFS';
-import { StateFile } from '@core/features/files/StateFile';
 import { TagsController } from '@core/features/tags/controller/TagsController';
 import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
 
-import { FlexSearchIndex } from '../../../database/flexsearch/FlexSearchIndex';
-
 import { NotesController } from './NotesController';
-import { NotesTextIndexer } from './NotesTextIndexer';
-
-const FAKE_WORKSPACE_ID = getUUID();
 
 describe('CRUD operations', () => {
 	const { getDB } = makeAutoClosedSQLiteDB();
+	const getAppContext = makeAppContext(getDB);
 
-	test('insertion multiple entries', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+	test('create few notes', async () => {
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		const entries = [
 			{ title: 'Title 1', text: 'Text 1' },
@@ -31,6 +23,7 @@ describe('CRUD operations', () => {
 		const ids = await Promise.all(entries.map((note) => registry.add(note)));
 		expect(ids).toHaveLength(entries.length);
 
+		// TODO: move to another test
 		// Entries match data
 		await registry.get().then((dbEntries) => {
 			dbEntries.forEach((dbEntry) => {
@@ -43,8 +36,8 @@ describe('CRUD operations', () => {
 	});
 
 	test('update entry and get by id', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		// Entries match data
 		const entries = await registry.get();
@@ -63,7 +56,11 @@ describe('CRUD operations', () => {
 	test('delete entries', async () => {
 		const dbFile = createFileControllerMock();
 		const db = await openSQLite(dbFile);
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		onTestFinished(() => db.close());
+
+		const workspaceId = await createWorkspaceId(db);
+
+		const registry = new NotesController(db, workspaceId);
 
 		// Insert entries to test
 		const notesSample = Array(300)
@@ -93,22 +90,54 @@ describe('CRUD operations', () => {
 				expect(length).toBe(lengthBeforeDelete - 1);
 			});
 		});
-
-		await db.close();
 	});
 
 	test('delete with empty list does not throw', async () => {
 		const dbFile = createFileControllerMock();
 		const db = await openSQLite(dbFile);
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		onTestFinished(() => db.close());
+		const workspaceId = await createWorkspaceId(db);
+
+		const registry = new NotesController(db, workspaceId);
 
 		await expect(registry.delete([])).resolves.not.toThrow();
 	});
 });
 
-describe('data fetching', () => {
-	const { getDB } = makeAutoClosedSQLiteDB();
+test('Many instances reads the data consistently', async () => {
+	const db = await openSQLite(createFileControllerMock());
+	onTestFinished(db.close);
 
+	const workspaceId = await createWorkspaceId(db);
+
+	// Add few notes via one controller
+	const controller1 = new NotesController(db, workspaceId);
+
+	const notes = [
+		{ title: 'Title 1', text: 'Text 1' },
+		{ title: 'Title 2', text: 'Text 2' },
+		{ title: 'Title 3', text: 'Text 3' },
+	];
+
+	await Promise.all(notes.map((note) => controller1.add(note)));
+
+	const notesList1 = await controller1.get();
+	expect(notesList1).toHaveLength(3);
+
+	// Read the state in another controller
+	const controller2 = new NotesController(db, workspaceId);
+
+	await expect(controller2.get()).resolves.toEqual(notesList1);
+});
+
+test('Get notes by pages', async () => {
+	const db = await openSQLite(createFileControllerMock());
+	onTestFinished(db.close);
+
+	const workspaceId = await createWorkspaceId(db);
+	const registry = new NotesController(db, workspaceId);
+
+	// Add data
 	const notesSample = Array(300)
 		.fill(null)
 		.map((_, idx) => {
@@ -118,299 +147,35 @@ describe('data fetching', () => {
 			};
 		});
 
-	test('insert sample entries', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+	const ids: string[] = [];
+	for (const note of notesSample) {
+		ids.push(await registry.add(note));
+	}
 
-		const ids: string[] = [];
-		for (const note of notesSample) {
-			ids.push(await registry.add(note));
-		}
+	const tags = new TagsController(db, workspaceId);
+	await tags.setAttachedTags(ids[0], [await tags.add('foo', null)]);
+	await tags.setAttachedTags(ids[1], [await tags.add('bar', null)]);
 
-		const tags = new TagsController(db, FAKE_WORKSPACE_ID);
-		await tags.setAttachedTags(ids[0], [await tags.add('foo', null)]);
-		await tags.setAttachedTags(ids[1], [await tags.add('bar', null)]);
-	});
+	await expect(registry.getLength()).resolves.toBe(notesSample.length);
 
-	test('filter by tags', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const tags = new TagsController(db, FAKE_WORKSPACE_ID);
+	// Invalid page must throw errors
+	await expect(registry.get({ limit: 100, page: 0 })).rejects.toThrow();
+	await expect(registry.get({ limit: 100, page: -100 })).rejects.toThrow();
 
-		const tagsList = await tags.getTags();
+	const page1 = await registry.get({ limit: 100, page: 1 });
+	expect(page1[0].content).toMatchObject(notesSample[0]);
 
-		await expect(
-			registry.get({
-				tags: [tagsList.find((tag) => tag.resolvedName === 'foo')!.id],
-			}),
-		).resolves.toHaveLength(1);
-
-		await expect(
-			registry.get({
-				tags: [tagsList.find((tag) => tag.resolvedName === 'bar')!.id],
-			}),
-		).resolves.toHaveLength(1);
-	});
-
-	test('get entries by pages', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		await registry.getLength().then((length) => {
-			expect(length).toBe(notesSample.length);
-		});
-
-		// Invalid page must throw errors
-		await expect(registry.get({ limit: 100, page: 0 })).rejects.toThrow();
-		await expect(registry.get({ limit: 100, page: -100 })).rejects.toThrow();
-
-		const page1 = await registry.get({ limit: 100, page: 1 });
-		expect(page1[0].content).toMatchObject(notesSample[0]);
-
-		const page2 = await registry.get({ limit: 100, page: 2 });
-		expect(page2[0].content).toMatchObject(notesSample[100]);
-	});
-
-	test('method query consider filters', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const tags = new TagsController(db, FAKE_WORKSPACE_ID);
-
-		await expect(registry.query()).resolves.toHaveLength(300);
-		await expect(registry.query({ limit: 100 })).resolves.toHaveLength(100);
-		await expect(
-			registry.query({ limit: 100, meta: { isDeleted: true } }),
-		).resolves.toHaveLength(0);
-
-		const tagsList = await tags.getTags();
-		const fooTag = tagsList.find((tag) => tag.resolvedName === 'foo');
-		expect(fooTag).toBeDefined();
-
-		await expect(
-			registry.query({
-				tags: [fooTag!.id],
-			}),
-		).resolves.toHaveLength(1);
-	});
-
-	test('filter by tags and the deleted status', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const tags = new TagsController(db, FAKE_WORKSPACE_ID);
-
-		const tagsList = await tags.getTags();
-
-		const barTag = tagsList.find((tag) => tag.resolvedName === 'bar')!.id;
-		const [{ id: noteId }] = await registry.get({
-			limit: 1,
-			tags: [barTag],
-		});
-
-		// set deleted status
-		await registry.updateMeta([noteId], { isDeleted: true });
-		await expect(
-			registry.get({
-				tags: [barTag],
-				meta: { isDeleted: true },
-			}),
-		).resolves.toHaveLength(1);
-
-		// reset deleted status
-		await registry.updateMeta([noteId], { isDeleted: false });
-		await expect(
-			registry.get({
-				tags: [barTag],
-				meta: { isDeleted: true },
-			}),
-		).resolves.toHaveLength(0);
-	});
-
-	test('get entries by deletion status', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const notesId = await registry
-			.get({ limit: 10 })
-			.then((notes) => notes.map((note) => note.id));
-
-		// update status for 10 notes
-		vi.setSystemTime('2020-01-01T15:00:00.000Z');
-		await registry.updateMeta(notesId, { isDeleted: true });
-		await expect(registry.get({ meta: { isDeleted: true } })).resolves.toHaveLength(
-			10,
-		);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					from: new Date('2020-01-01T15:00:00.000Z'),
-					to: new Date('2020-01-01T15:00:00.000Z'),
-				},
-			}),
-			'Must find by exact time',
-		).resolves.toHaveLength(10);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					from: new Date('2020-01-01T15:00:00.000Z'),
-				},
-			}),
-			'Must find by exact "from" time',
-		).resolves.toHaveLength(10);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					from: new Date('2020-01-01T14:00:00.000Z'),
-				},
-			}),
-			'Must find by "from" time before deletion',
-		).resolves.toHaveLength(10);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					from: new Date('2020-01-01T15:00:00.001Z'),
-				},
-			}),
-			'Must not find by "from" time after deletion',
-		).resolves.toHaveLength(0);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					to: new Date('2020-01-01T15:00:00.000Z'),
-				},
-			}),
-			'Must find by exact "to" time',
-		).resolves.toHaveLength(10);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					to: new Date('2020-01-01T15:00:00.001Z'),
-				},
-			}),
-			'Must find by "to" time after deletion',
-		).resolves.toHaveLength(10);
-
-		await expect(
-			registry.get({
-				deletedAt: {
-					to: new Date('2020-01-01T14:00:00.000Z'),
-				},
-			}),
-			'Must not find by "to" time before deletion',
-		).resolves.toHaveLength(0);
-
-		// check only not deleted notes
-		await expect(registry.get({ meta: { isDeleted: false } })).resolves.toHaveLength(
-			notesSample.length - 10,
-		);
-	});
-
-	test('filter notes by archived status', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const notesId = await registry
-			.get({ limit: 30 })
-			.then((notes) => notes.map((note) => note.id));
-
-		await registry.updateMeta(notesId, { isArchived: true });
-		await expect(registry.get({ meta: { isArchived: true } })).resolves.toHaveLength(
-			30,
-		);
-		await expect(registry.get({ meta: { isArchived: false } })).resolves.toHaveLength(
-			notesSample.length - 30,
-		);
-	});
-
-	test('filters notes by bookmarks', async () => {
-		const db = await getDB();
-
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const notesId = await registry
-			.get({ limit: 40 })
-			.then((notes) => notes.map((note) => note.id));
-
-		await registry.updateMeta(notesId, { isBookmarked: true });
-		await expect(
-			registry.get({ meta: { isBookmarked: true } }),
-		).resolves.toHaveLength(40);
-		await expect(
-			registry.get({ meta: { isBookmarked: false } }),
-		).resolves.toHaveLength(notesSample.length - 40);
-	});
-
-	test('method getLength consider filters', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const notesId = await registry
-			.get({ limit: 10 })
-			.then((notes) => notes.map((note) => note.id));
-		await registry.updateMeta(notesId, { isVisible: false });
-
-		await expect(registry.getLength({ meta: { isVisible: false } })).resolves.toBe(
-			10,
-		);
-		await expect(registry.getLength({ meta: { isVisible: true } })).resolves.toBe(
-			notesSample.length - 10,
-		);
-	});
-
-	test('method getById returns notes in the requested order', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const notes = await registry.get({ limit: 10 });
-
-		const shuffledNotes = notes.sort(() => Math.random() - 0.5);
-		const shuffledIds = shuffledNotes.map((n) => n.id);
-
-		await expect(registry.getById(shuffledIds)).resolves.toEqual(shuffledNotes);
-	});
+	const page2 = await registry.get({ limit: 100, page: 2 });
+	expect(page2[0].content).toMatchObject(notesSample[100]);
 });
 
-describe('multi instances', () => {
-	const { getDB } = makeAutoClosedSQLiteDB({ closeHook: afterEach });
-
-	test('insertion multiple entries and close with sync data', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const notes = [
-			{ title: 'Title 1', text: 'Text 1' },
-			{ title: 'Title 2', text: 'Text 2' },
-			{ title: 'Title 3', text: 'Text 3' },
-		];
-
-		await Promise.all(notes.map((note) => registry.add(note)));
-		await db.sync();
-	});
-
-	test('read entries from previous step', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		// Entries match data
-		const entries = await registry.get();
-
-		expect(entries).toHaveLength(3);
-		expect(entries[1].content.title.length).toBeGreaterThan(0);
-		expect(entries[1].content.text.length).toBeGreaterThan(0);
-	});
-});
-
-describe('Notes meta control', () => {
+describe('Meta data control', () => {
 	const { getDB } = makeAutoClosedSQLiteDB();
+	const getAppContext = makeAppContext(getDB);
 
 	test('toggle note versions control', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		// Create note
 		const noteId = await registry.add({ title: 'Title', text: 'Text' });
@@ -441,8 +206,8 @@ describe('Notes meta control', () => {
 	});
 
 	test('toggle note visibility', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		// Create note
 		const noteId = await registry.add({ title: 'Title', text: 'Text' });
@@ -488,8 +253,8 @@ describe('Notes meta control', () => {
 	});
 
 	test('toggle note deletion status', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		// Create notes
 		const noteId = await registry.add({ title: 'Title', text: 'Text' });
@@ -520,8 +285,8 @@ describe('Notes meta control', () => {
 	});
 
 	test('toggle note archived status', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		// Create notes
 		const noteId = await registry.add({ title: 'Title', text: 'Text' });
@@ -552,8 +317,8 @@ describe('Notes meta control', () => {
 	});
 
 	test('toggle note bookmarked status', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		const noteId = await registry.add({ title: 'Title', text: 'Text' });
 		await expect(registry.getById([noteId])).resolves.toMatchObject([
@@ -577,190 +342,6 @@ describe('Notes meta control', () => {
 				id: noteId,
 				isBookmarked: false,
 			},
-		]);
-	});
-});
-
-describe('Notes search', () => {
-	const { getDB } = makeAutoClosedSQLiteDB();
-
-	const index = new FlexSearchIndex(new InMemoryFS());
-
-	const texts = [
-		'A fast auburn fox leaped above a sleepy canine',
-		'PostgreSQL trigram indexing speeds up fuzzy text search',
-		'Full-text search and trigram search solve different problems',
-		'The database index makes LIKE queries fast',
-		'Fuzzy string matching can catch small typos in text',
-		'A lazy dog was sleeping near the river bank',
-		'The quick red fox jumped high above the hedge',
-		'Speedy foxes and sleepy dogs appear in many idioms',
-		'The quick brown fox jumps over the lazy dog',
-		'Trigram search helps find similar product names quickly',
-	];
-
-	let tagsMap: Record<string, string>;
-	test('Add notes', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const tags = new TagsController(db, FAKE_WORKSPACE_ID);
-
-		tagsMap = await Promise.all(
-			['1th', '2th', '3th', 'none'].map(async (name) => [
-				name,
-				await tags.add(name, null),
-			]),
-		).then((tags) => Object.fromEntries(tags));
-
-		await Promise.all(
-			texts.map(async (text, index) => {
-				const noteId = await registry.add({ title: `Note #${index + 1}`, text });
-
-				const seqNum = index + 1;
-				await tags.setAttachedTags(
-					noteId,
-					[
-						tagsMap['1th'],
-						seqNum % 2 === 0 && tagsMap['2th'],
-						seqNum % 3 === 0 && tagsMap['3th'],
-					].filter(Boolean) as string[],
-				);
-
-				return noteId;
-			}),
-		);
-	});
-
-	test('Update lexemes', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-
-		const indexScannerFile = createFileControllerMock();
-		const indexScanner = new NotesTextIndexer(
-			registry,
-			index,
-			new StateFile(indexScannerFile, z.any()),
-		);
-		await expect(indexScanner.update()).resolves.toBeGreaterThan(0);
-	});
-
-	test('Search by text', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID, index);
-
-		await expect(
-			registry.get({ search: { text: 'fox' }, limit: 3 }),
-			'Top results includes a word for search',
-		).resolves.toEqual(
-			Array.from({ length: 3 }).map(() =>
-				expect.objectContaining({
-					content: expect.objectContaining({
-						text: expect.stringContaining('fox'),
-					}),
-				}),
-			),
-		);
-
-		await expect(
-			registry.get({ search: { text: 'FOX' }, limit: 3 }),
-			'Search ignores text case',
-		).resolves.toEqual(
-			Array.from({ length: 3 }).map(() =>
-				expect.objectContaining({
-					content: expect.objectContaining({
-						text: expect.stringContaining('fox'),
-					}),
-				}),
-			),
-		);
-
-		await expect(
-			registry.get({ search: { text: 'qick' }, limit: 3 }),
-			'Search is typo tolerant',
-		).resolves.toEqual(
-			Array.from({ length: 2 }).map(() =>
-				expect.objectContaining({
-					content: expect.objectContaining({
-						text: expect.stringContaining('quick'),
-					}),
-				}),
-			),
-		);
-
-		await expect(
-			registry.get({ search: { text: '#4' }, limit: 1 }),
-			'Search works for note titles',
-		).resolves.toEqual([
-			expect.objectContaining({
-				content: expect.objectContaining({
-					title: 'Note #4',
-				}),
-			}),
-		]);
-	});
-
-	test('Search by text and filter by tags', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID, index);
-
-		await expect(
-			registry.get({ search: { text: 'Note' } }),
-			'Fetch notes for control',
-		).resolves.toHaveLength(texts.length);
-
-		await expect(
-			registry.get({ search: { text: 'Note' }, meta: { isVisible: false } }),
-			'Filter by meta properties hides results',
-		).resolves.toHaveLength(0);
-
-		await expect(
-			registry.get({ search: { text: 'Note' }, tags: [tagsMap['none']] }),
-			'Filter by not attached tag returns no results',
-		).resolves.toHaveLength(0);
-
-		await expect(
-			registry.get({ search: { text: 'Note' }, tags: [tagsMap['1th']] }),
-			'Filter by specific tag',
-		).resolves.toHaveLength(texts.length);
-
-		await expect(
-			registry.get({ search: { text: 'Note' }, tags: [tagsMap['2th']] }),
-			'Filter by specific tag',
-		).resolves.toHaveLength(Math.floor(texts.length / 2));
-
-		await expect(
-			registry.get({ search: { text: 'Note' }, tags: [tagsMap['3th']] }),
-			'Filter by specific tag',
-		).resolves.toHaveLength(Math.floor(texts.length / 3));
-
-		await expect(
-			registry.get({
-				search: { text: 'Note #2' },
-				tags: [tagsMap['3th']],
-				limit: 1,
-			}),
-			'Search returns most similar text after filtering',
-		).resolves.not.toEqual([
-			expect.objectContaining({
-				content: expect.objectContaining({
-					title: 'Note #2',
-				}),
-			}),
-		]);
-
-		await expect(
-			registry.get({
-				search: { text: 'Note #2' },
-				tags: [tagsMap['2th']],
-				limit: 1,
-			}),
-			'Exact match with filter must be at top of the list',
-		).resolves.toEqual([
-			expect.objectContaining({
-				content: expect.objectContaining({
-					title: 'Note #2',
-				}),
-			}),
 		]);
 	});
 });
