+++ src/core/features/notes/history/NoteVersions.test.ts
@@ -1,18 +1,17 @@
+import { makeAppContext } from 'src/__tests__/utils/makeAppContext';
 import { makeAutoClosedSQLiteDB } from 'src/__tests__/utils/makeAutoClosedSQLiteDB';
-import { getUUID } from 'src/__tests__/utils/uuid';
 
 import { NotesController } from '../controller/NotesController';
 import { NoteVersions } from './NoteVersions';
 
-const FAKE_WORKSPACE_ID = getUUID();
-
 describe('Note version control', () => {
 	const { getDB } = makeAutoClosedSQLiteDB();
+	const getAppContext = makeAppContext(getDB);
 
 	test('snapshot must be created only if latest version have changes with latest note data', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const history = new NoteVersions(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const history = new NoteVersions(db, workspaceId);
 
 		// Fetch note
 		const dataV1 = { title: 'Title', text: 'Text' };
@@ -51,9 +50,9 @@ describe('Note version control', () => {
 	});
 
 	test('snapshot must always be created when parameter `force` is set', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const history = new NoteVersions(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const history = new NoteVersions(db, workspaceId);
 
 		// Fetch note
 		const noteId = await registry.add({ title: 'Title 1', text: 'Text 1' });
@@ -96,9 +95,9 @@ describe('Note version control', () => {
 	});
 
 	test('when note is deleted, all versions must be deleted too', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const history = new NoteVersions(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const history = new NoteVersions(db, workspaceId);
 
 		// Fetch note
 		const noteId = await registry.add({ title: 'Title 1', text: 'Text 1' });
@@ -120,9 +119,9 @@ describe('Note version control', () => {
 	});
 
 	test('snapshot may be created with empty text', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const history = new NoteVersions(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const history = new NoteVersions(db, workspaceId);
 
 		// Fetch note
 		const dataV1 = { title: '', text: '' };
@@ -167,10 +166,11 @@ describe('Note version control', () => {
 
 describe('Delete note versions', () => {
 	const { getDB } = makeAutoClosedSQLiteDB();
+	const getAppContext = makeAppContext(getDB);
 
 	test('create few notes', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
 
 		await Promise.all(
 			[
@@ -182,9 +182,9 @@ describe('Delete note versions', () => {
 	});
 
 	test('purge all versions for note', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const history = new NoteVersions(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const history = new NoteVersions(db, workspaceId);
 
 		// Fetch any note
 		const notes = await registry.get();
@@ -206,9 +206,9 @@ describe('Delete note versions', () => {
 	});
 
 	test('delete specific versions for note', async () => {
-		const db = await getDB();
-		const registry = new NotesController(db, FAKE_WORKSPACE_ID);
-		const history = new NoteVersions(db, FAKE_WORKSPACE_ID);
+		const { db, workspaceId } = getAppContext();
+		const registry = new NotesController(db, workspaceId);
+		const history = new NoteVersions(db, workspaceId);
 
 		// Fetch any note
 		const notes = await registry.get();
